core: stdcm: move dummy block occupancy to test modules
