+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/preprocessing/OccupancySegment.kt
@@ -1,4 +1,4 @@
-package fr.sncf.osrd.stdcm
+package fr.sncf.osrd.stdcm.preprocessing
 
 import fr.sncf.osrd.sim_infra.api.BlockId
 import fr.sncf.osrd.utils.units.Distance
@@ -7,7 +7,6 @@ import fr.sncf.osrd.utils.units.Distance
  * The given element is unavailable from timeStart until timeEnd, in the space between distanceStart
  * and distanceEnd. Distances are relative to the start of the element.
  */
-@JvmRecord
 data class OccupancySegment(
     val timeStart: Double,
     val timeEnd: Double,
